Add adding profiles in database

+++ GGenToPrint/Resources/Drawables/Sheet/SheetDrawable.cs
@@ -16,35 +16,35 @@
             canvas.StrokeColor = drawColor;
 
             // Get cell size
-            float cellSize = rectF.Width / (float)NumCellsOfHorizontal;
+            float cellSize = rectF.Width / NumCellsOfHorizontal;
             if (NumCellsOfVertical * cellSize > rectF.Height)
             {
-                cellSize = rectF.Height / (float)NumCellsOfVertical;
+                cellSize = rectF.Height / NumCellsOfVertical;
             }
 
             // Borders
-            canvas.DrawLine(rectF.Left, rectF.Top, cellSize * (float)NumCellsOfHorizontal, rectF.Top);
-            canvas.DrawLine(cellSize * (float)NumCellsOfHorizontal, rectF.Top, cellSize * (float)NumCellsOfHorizontal, cellSize * (float)NumCellsOfVertical);
-            canvas.DrawLine(cellSize * (float)NumCellsOfHorizontal, cellSize * (float)NumCellsOfVertical, rectF.Left, cellSize * (float)NumCellsOfVertical);
-            canvas.DrawLine(rectF.Left, cellSize * (float)NumCellsOfVertical, rectF.Left, rectF.Top);
+            canvas.DrawLine(rectF.Left, rectF.Top, cellSize * NumCellsOfHorizontal, rectF.Top);
+            canvas.DrawLine(cellSize * NumCellsOfHorizontal, rectF.Top, cellSize * NumCellsOfHorizontal, cellSize * NumCellsOfVertical);
+            canvas.DrawLine(cellSize * NumCellsOfHorizontal, cellSize * NumCellsOfVertical, rectF.Left, cellSize * NumCellsOfVertical);
+            canvas.DrawLine(rectF.Left, cellSize * NumCellsOfVertical, rectF.Left, rectF.Top);
 
             // Cells and rows
             if (SheetTypeIndex == 0)
             {
                 for (int i = 1; i < NumCellsOfHorizontal; i++)
                 {
-                    canvas.DrawLine(cellSize * i, rectF.Top, cellSize * i, cellSize * (float)NumCellsOfVertical);
+                    canvas.DrawLine(cellSize * i, rectF.Top, cellSize * i, cellSize * NumCellsOfVertical);
                 }
                 for (int i = 1; i < NumCellsOfVertical; i++)
                 {
-                    canvas.DrawLine(rectF.Left, cellSize * i, cellSize * (float)NumCellsOfHorizontal, cellSize * i);
+                    canvas.DrawLine(rectF.Left, cellSize * i, cellSize * NumCellsOfHorizontal, cellSize * i);
                 }
             }
             else if (SheetTypeIndex == 1)
             {
                 for (int i = 1; i < NumCellsOfVertical; i++)
                 {
-                    canvas.DrawLine(rectF.Left, cellSize * i, cellSize * (float)NumCellsOfHorizontal, cellSize * i);
+                    canvas.DrawLine(rectF.Left, cellSize * i, cellSize * NumCellsOfHorizontal, cellSize * i);
                 }
             }
 
@@ -56,18 +56,18 @@
                 if (SheetPositionIndex == 0)
                 {
                     canvas.DrawLine(
-                        cellSize * (float)NumCellsOfHorizontal - cellSize * (float)NumCellsOfMargin,
+                        cellSize * NumCellsOfHorizontal - cellSize * NumCellsOfMargin,
                         rectF.Top,
-                        cellSize * (float)NumCellsOfHorizontal - cellSize * (float)NumCellsOfMargin,
-                        cellSize * (float)NumCellsOfVertical);
+                        cellSize * NumCellsOfHorizontal - cellSize * NumCellsOfMargin,
+                        cellSize * NumCellsOfVertical);
                 }
                 else if (SheetPositionIndex == 1)
                 {
                     canvas.DrawLine(
-                        cellSize * (float)NumCellsOfMargin,
+                        cellSize * NumCellsOfMargin,
                         rectF.Top,
-                        cellSize * (float)NumCellsOfMargin,
-                        cellSize * (float)NumCellsOfVertical);
+                        cellSize * NumCellsOfMargin,
+                        cellSize * NumCellsOfVertical);
                 }
                 canvas.StrokeSize = 1;
                 canvas.StrokeColor = drawColor;
